requirements.txt
websocket-client==1.9.0
sherpa-onnx==1.12.34
sounddevice==0.5.5
numpy==2.4.3
edge-tts==7.2.8
faster-whisper==1.2.1
